requirements.txt
distlib==0.3.9
dnspython==2.7.0
filelock==3.17.0
getmac==0.9.5
platformdirs==4.3.6
pymongo==4.11.1
python-dotenv==1.0.1
virtualenv==20.29.2
